scripts/requirements.txt
# PPO drone control scripts (train.py, val.py)
# Also need: flightlib (flightgym) from flightmare/flightlib with FLIGHTMARE_PATH set
gymnasium>=0.29.0
numpy>=1.20.0
PyYAML>=5.4
stable-baselines3>=2.0.0
matplotlib>=3.5.0
